+++ examples/vue3/src/App.vue
@@ -1,9 +1,8 @@
 <script setup lang="ts">
-import { useOpenModal, useCloseModal, useUpdateModal } from '@usemodals/vue3'
-import { Modal, Button } from 'ant-design-vue'
+import { useOpenModal, useUpdateModal } from '@usemodals/vue3'
+import { Button } from 'ant-design-vue'
 
 const openModal = useOpenModal()
-const closeModal = useCloseModal()
 const updateModal = useUpdateModal()
 
 const openAndUpdate = (merge?: boolean) => {
@@ -14,21 +13,15 @@ const openAndUpdate = (merge?: boolean) => {
 }
 
 import Demo from './demo'
+import Test from './test.vue'
 </script>
 
 <template>
-  <UseModal modalId="modalId">
-    <template #modal="modalState">
-      <Modal :visible="modalState.visible" @cancel="closeModal('modalId')">
-        {{ modalState.props }}
-      </Modal>
-    </template>
-  </UseModal>
-
   <Button @click="openModal('modalId', { title: 'MODAL TITLE' })">open modal</Button>
   <Button @click="() => openAndUpdate()">open modal and update props after 3s(override props)</Button>
   <Button @click="openAndUpdate(true)">open modal and update props after 3s(merge props)</Button>
 
   <Demo />
+  <Test />
 </template>