+++ examples/vue3/src/demo.tsx
@@ -1,6 +1,8 @@
 import { defineComponent } from 'vue'
-import { Button, Modal } from 'ant-design-vue'
+import { Button } from 'ant-design-vue'
 import { UseModal, useCloseModal, useOpenModal, useUpdateModal } from '@usemodals/vue3'
+
+import Modal from './modal'
 
 export default defineComponent({
   setup() {
@@ -17,16 +19,7 @@ export default defineComponent({
 
     return () => (
       <div>
-        <UseModal modalId="modalIdDemo">
-          {{
-            modal: (modalState) => (
-              <Modal visible={modalState.visible} title={modalState.props.title} onCancel={() => closeModal('modalIdDemo')}>
-                { modalState.props.content }
-              </Modal>
-            )
-          }}
-        </UseModal>
-
+        <Modal />
         <Button onClick={() => openModal('modalIdDemo', { title: 'MODAL TITLE' })}>demo open modal</Button>
         <Button onClick={() => openAndUpdate()}>demo open modal and update props after 3s(override props)</Button>
         <Button onClick={() => openAndUpdate(true)}>demo open modal and update props after 3s(merge props)</Button>