+++ packages/react/src/index.ts
@@ -9,5 +9,6 @@ export { useCloseModal } from './hooks/useCloseModal'
 export { useUpdateModal } from './hooks/useUpdateModal'
 export { useModalIsLoading } from './hooks/useModalIsLoading'
 export { useRegisterModal } from './hooks/useRegisterModal'
+export { useModalProps } from './hooks/useModalProps'
 
 export type { ModalBasicProps } from './types'