+++ packages/vue3/src/index.ts
@@ -1,6 +1,6 @@
 import { UseModalsPlugin } from './plugin'
 
-export { useOpenModal, useUpdateModal, useCloseModal, useCloseAllModals } from './hooks'
+export { useOpenModal, useUpdateModal, useCloseModal, useCloseAllModals, useModalProps } from './hooks'
 export { UseModal } from './use-modal'
 
 export default UseModalsPlugin