+++ packages/react/CHANGELOG.md
@@ -9,3 +9,11 @@
 - Implement `useUpdateModal` function
 - Implement `useCloseModal` function
 - Implement `useCloseAllModals` function
+
+## [@usemodals/react@1.1.0] - 2023-06-01
+
+### <Features>
+- Implement `useModalProps` function
+
+### <Changed>
+- before update Modal props, check it's visible